pull types and helper functions out of the menu file
